REFACTOR : src/lib/hooks/api.ts=>src/apis/apis.ts; src/lib/client.ts=>src/apis/client.ts
